refactor(edge): extract h3 bridge protocol tests
