+++ crates/edge/tests/h3_bridge.rs
@@ -39,6 +39,8 @@ use tokio_rustls::{
     rustls::{ClientConfig, RootCertStore, pki_types::ServerName},
 };
 
+mod support;
+
 use spooky_config::config::{
     Backend, ClientAuth, Config, HealthCheck, Listen, LoadBalancing, Log, LogFormat, Security, Tls,
     TlsCertificate, UpstreamTls,
@@ -51,30 +53,11 @@ use spooky_edge::constants::{
     QUIC_INITIAL_MAX_STREAMS_BIDI, QUIC_INITIAL_MAX_STREAMS_UNI, QUIC_INITIAL_STREAM_DATA,
     REQUEST_TIMEOUT_SECS, UDP_READ_TIMEOUT_MS,
 };
+use support::net::local_listener_bind_available;
 
 type TrailerPairs = Vec<(String, String)>;
 type H3TrailerResponse = (String, Vec<u8>, TrailerPairs);
 type BootstrapResponse = (StatusCode, Vec<u8>, TrailerPairs);
-
-fn local_listener_bind_available() -> bool {
-    let tcp_ok = match std::net::TcpListener::bind("127.0.0.1:0") {
-        Ok(listener) => {
-            drop(listener);
-            true
-        }
-        Err(err) if err.kind() == std::io::ErrorKind::PermissionDenied => false,
-        Err(_) => true,
-    };
-    let udp_ok = match std::net::UdpSocket::bind("127.0.0.1:0") {
-        Ok(socket) => {
-            drop(socket);
-            true
-        }
-        Err(err) if err.kind() == std::io::ErrorKind::PermissionDenied => false,
-        Err(_) => true,
-    };
-    tcp_ok && udp_ok
-}
 
 fn write_test_certs(dir: &TempDir) -> (String, String) {
     write_named_test_cert(dir, "cert", &["localhost"], &[IpAddr::from([127, 0, 0, 1])])
@@ -2321,716 +2304,8 @@ fn make_config_with_backends(
     config
 }
 
-#[test]
-fn http3_to_http2_roundtrip() {
-    if !local_listener_bind_available() {
-        return;
-    }
-    let dir = tempdir().expect("failed to create temp dir");
-    let (cert, key) = write_test_certs(&dir);
-
-    let rt = tokio::runtime::Runtime::new().expect("runtime");
-    let backend_addr = rt.block_on(start_h2_backend_with_regression_routes());
-    let config = make_config(0, backend_addr.to_string(), cert, key);
-    let listener = QUICListener::new(config).expect("failed to create listener");
-    let listen_addr = listener.socket.local_addr().unwrap();
-    let _listener_task = ListenerTaskGuard::spawn(&rt, listener);
-
-    let body = run_h3_client(listen_addr).expect("client request failed");
-
-    assert!(!body.is_empty(), "expected non-empty response from backend");
-}
-
-#[test]
-fn http3_to_http2_preserves_response_trailers() {
-    if !local_listener_bind_available() {
-        return;
-    }
-    let dir = tempdir().expect("failed to create temp dir");
-    let (cert, key) = write_test_certs(&dir);
-
-    let rt = tokio::runtime::Runtime::new().expect("runtime");
-    let backend_addr = rt.block_on(start_h2_backend_with_trailers());
-    let config = make_config(0, backend_addr.to_string(), cert, key);
-    let listener = QUICListener::new(config).expect("failed to create listener");
-    let listen_addr = listener.socket.local_addr().unwrap();
-    let _listener_task = ListenerTaskGuard::spawn(&rt, listener);
-
-    let (status, body, trailers) =
-        run_h3_client_collect_trailers(listen_addr, "/").expect("client trailer request failed");
-
-    assert_eq!(status, "200");
-    assert_eq!(body, b"hello\n");
-    assert!(
-        trailers
-            .iter()
-            .any(|(name, value)| name.eq_ignore_ascii_case("grpc-status") && value == "0"),
-        "expected grpc-status trailer, got {trailers:?}"
-    );
-    assert!(
-        trailers
-            .iter()
-            .any(|(name, value)| name.eq_ignore_ascii_case("grpc-message") && value == "ok"),
-        "expected grpc-message trailer, got {trailers:?}"
-    );
-}
-
-#[test]
-fn bootstrap_h2_preserves_response_trailers() {
-    if !local_listener_bind_available() {
-        return;
-    }
-    let dir = tempdir().expect("failed to create temp dir");
-    let (cert, key) = write_test_certs(&dir);
-
-    let rt = tokio::runtime::Runtime::new().expect("runtime");
-    let backend_addr = rt.block_on(start_h2_backend_with_trailers());
-    let listen_port = find_free_tcp_port();
-    let config = make_config(
-        u32::from(listen_port),
-        backend_addr.to_string(),
-        cert.clone(),
-        key,
-    );
-    let listener = make_listener_with_bootstrap(config);
-    let listen_addr = listener.socket.local_addr().unwrap();
-    let _listener_task = ListenerTaskGuard::spawn(&rt, listener);
-
-    let bootstrap_addr = SocketAddr::new(listen_addr.ip(), listen_port);
-    let (status, body, trailers) = rt
-        .block_on(run_bootstrap_h2_client_collect_trailers(
-            bootstrap_addr,
-            &cert,
-            "/",
-        ))
-        .expect("bootstrap trailer request failed");
-
-    assert_eq!(status, StatusCode::OK);
-    assert_eq!(body, b"hello\n");
-    assert!(
-        trailers
-            .iter()
-            .any(|(name, value)| name.eq_ignore_ascii_case("grpc-status") && value == "0"),
-        "expected grpc-status trailer, got {trailers:?}"
-    );
-    assert!(
-        trailers
-            .iter()
-            .any(|(name, value)| name.eq_ignore_ascii_case("grpc-message") && value == "ok"),
-        "expected grpc-message trailer, got {trailers:?}"
-    );
-}
-
-#[test]
-#[serial_test::serial]
-fn http3_sni_selects_exact_and_fallback_certificates_on_real_handshake() {
-    if !local_listener_bind_available() {
-        return;
-    }
-    let dir = tempdir().expect("failed to create temp dir");
-    let (fallback_cert, fallback_key) =
-        write_named_test_cert(&dir, "fallback", &["fallback.example.com"], &[]);
-    let (api_cert, api_key) = write_named_test_cert(&dir, "api", &["api.example.com"], &[]);
-    let expected_api_der = read_test_leaf_der(&api_cert).expect("api cert der");
-    let expected_fallback_der = read_test_leaf_der(&fallback_cert).expect("fallback cert der");
-    let rt = tokio::runtime::Runtime::new().expect("runtime");
-    let backend_addr = rt.block_on(start_h2_backend());
-
-    let listen_port = find_free_tcp_port();
-    let mut config = make_config(
-        listen_port as u32,
-        backend_addr.to_string(),
-        fallback_cert.clone(),
-        fallback_key,
-    );
-    config.listen.tls.certificates = vec![TlsCertificate {
-        server_name: "api.example.com".to_string(),
-        cert: api_cert.clone(),
-        key: api_key,
-    }];
-
-    let _enter = rt.enter();
-    let listener = make_listener_with_bootstrap(config);
-    drop(_enter);
-    let listen_addr = listener.socket.local_addr().expect("listener addr");
-    let _listener_task = ListenerTaskGuard::spawn(&rt, listener);
-
-    let exact = run_h3_client_with_tls(
-        listen_addr,
-        H3TlsClientOptions {
-            server_name: "api.example.com",
-            authority: "api.example.com",
-            path: "/",
-            verify_peer: false,
-            root_cert_path: None,
-            client_identity: None,
-            application_protos: quiche::h3::APPLICATION_PROTOCOL,
-            send_request: true,
-        },
-    )
-    .expect("exact sni client request failed");
-    assert_eq!(exact.alpn, b"h3");
-    assert_eq!(
-        exact.peer_cert.as_deref(),
-        Some(expected_api_der.as_slice()),
-        "exact SNI should select api.example.com certificate"
-    );
-    assert_eq!(String::from_utf8_lossy(&exact.body), "backend ok\n");
-
-    let fallback = run_h3_client_with_tls(
-        listen_addr,
-        H3TlsClientOptions {
-            server_name: "other.example.com",
-            authority: "other.example.com",
-            path: "/",
-            verify_peer: false,
-            root_cert_path: None,
-            client_identity: None,
-            application_protos: quiche::h3::APPLICATION_PROTOCOL,
-            send_request: true,
-        },
-    )
-    .expect("fallback sni client request failed");
-    assert_eq!(fallback.alpn, b"h3");
-    assert_eq!(
-        fallback.peer_cert.as_deref(),
-        Some(expected_fallback_der.as_slice()),
-        "unmatched SNI should fall back to default certificate"
-    );
-    assert_eq!(String::from_utf8_lossy(&fallback.body), "backend ok\n");
-}
-
-#[test]
-#[serial_test::serial]
-fn bootstrap_h2_optional_client_auth_allows_requests_without_certificate() {
-    if !local_listener_bind_available() {
-        return;
-    }
-    let dir = tempdir().expect("failed to create temp dir");
-    let (cert, key) = write_test_certs(&dir);
-    let (ca_cert, _client_cert, _client_key) =
-        write_test_ca_and_client_cert(&dir, "client-ca", "client.example.com");
-    let rt = tokio::runtime::Runtime::new().expect("runtime");
-    let backend_addr = rt.block_on(start_h2_backend());
-
-    let listen_port = find_free_tcp_port();
-    let mut config = make_config(
-        listen_port as u32,
-        backend_addr.to_string(),
-        cert.clone(),
-        key,
-    );
-    config.listen.tls.client_auth = ClientAuth {
-        enabled: true,
-        require_client_cert: false,
-        ca_file: Some(ca_cert),
-    };
-
-    let _enter = rt.enter();
-    let listener = make_listener_with_bootstrap(config);
-    drop(_enter);
-    let listen_addr = listener.socket.local_addr().expect("listener addr");
-    let _listener_task = ListenerTaskGuard::spawn(&rt, listener);
-    let bootstrap_addr = SocketAddr::new(listen_addr.ip(), listen_port);
-
-    let response = rt
-        .block_on(run_bootstrap_h2_client_request(
-            bootstrap_addr,
-            &cert,
-            "GET",
-            "/",
-            &[],
-        ))
-        .expect("optional client-auth request failed");
-    assert_eq!(response.0, StatusCode::OK);
-    assert_eq!(String::from_utf8_lossy(&response.1), "backend ok\n");
-}
-
-#[test]
-#[serial_test::serial]
-fn bootstrap_h2_required_client_auth_rejects_missing_certificate_and_accepts_valid_certificate() {
-    if !local_listener_bind_available() {
-        return;
-    }
-    let dir = tempdir().expect("failed to create temp dir");
-    let (cert, key) = write_test_certs(&dir);
-    let (ca_cert, client_cert, client_key) =
-        write_test_ca_and_client_cert(&dir, "client-ca", "client.example.com");
-    let rt = tokio::runtime::Runtime::new().expect("runtime");
-    let backend_addr = rt.block_on(start_h2_backend());
-
-    let listen_port = find_free_tcp_port();
-    let mut config = make_config(
-        listen_port as u32,
-        backend_addr.to_string(),
-        cert.clone(),
-        key,
-    );
-    config.listen.tls.client_auth = ClientAuth {
-        enabled: true,
-        require_client_cert: true,
-        ca_file: Some(ca_cert.clone()),
-    };
-
-    let _enter = rt.enter();
-    let listener = make_listener_with_bootstrap(config);
-    drop(_enter);
-    let listen_addr = listener.socket.local_addr().expect("listener addr");
-    let _listener_task = ListenerTaskGuard::spawn(&rt, listener);
-    let bootstrap_addr = SocketAddr::new(listen_addr.ip(), listen_port);
-
-    let missing_cert_err = rt
-        .block_on(run_bootstrap_h2_client_request_with_client_auth(
-            bootstrap_addr,
-            &cert,
-            "GET",
-            "/",
-            &[],
-            None,
-        ))
-        .expect_err("missing client certificate should fail");
-    assert!(
-        missing_cert_err.contains("tls connect")
-            || missing_cert_err.contains("sender ready")
-            || missing_cert_err.contains("send request"),
-        "unexpected missing-client-cert error: {missing_cert_err}"
-    );
-
-    let response = rt
-        .block_on(run_bootstrap_h2_client_request_with_client_auth(
-            bootstrap_addr,
-            &cert,
-            "GET",
-            "/",
-            &[],
-            Some((&client_cert, &client_key)),
-        ))
-        .expect("client-authenticated bootstrap request failed");
-    assert_eq!(response.0, StatusCode::OK);
-    assert_eq!(String::from_utf8_lossy(&response.1), "backend ok\n");
-}
-
-#[test]
-#[serial_test::serial]
-fn quic_tls_metrics_capture_selection_and_failures() {
-    if !local_listener_bind_available() {
-        return;
-    }
-    let dir = tempdir().expect("failed to create temp dir");
-    let (fallback_cert, fallback_key) =
-        write_named_test_cert(&dir, "fallback", &["fallback.example.com"], &[]);
-    let (api_cert, api_key) = write_named_test_cert(&dir, "api", &["api.example.com"], &[]);
-    let rt = tokio::runtime::Runtime::new().expect("runtime");
-    let backend_addr = rt.block_on(start_h2_backend());
-
-    let listen_port = find_free_tcp_port();
-    let metrics_port = find_free_tcp_port();
-    let mut config = make_config(
-        listen_port as u32,
-        backend_addr.to_string(),
-        fallback_cert.clone(),
-        fallback_key,
-    );
-    config.listen.tls.certificates = vec![TlsCertificate {
-        server_name: "api.example.com".to_string(),
-        cert: api_cert,
-        key: api_key,
-    }];
-    config.observability.metrics.enabled = true;
-    config.observability.metrics.address = "127.0.0.1".to_string();
-    config.observability.metrics.port = metrics_port;
-    config.observability.metrics.path = "/metrics".to_string();
-
-    let _enter = rt.enter();
-    let listener = make_listener_with_bootstrap(config);
-    drop(_enter);
-    let listen_addr = listener.socket.local_addr().expect("listener addr");
-    let _listener_task = ListenerTaskGuard::spawn(&rt, listener);
-
-    run_h3_client_with_tls(
-        listen_addr,
-        H3TlsClientOptions {
-            server_name: "api.example.com",
-            authority: "api.example.com",
-            path: "/",
-            verify_peer: false,
-            root_cert_path: None,
-            client_identity: None,
-            application_protos: quiche::h3::APPLICATION_PROTOCOL,
-            send_request: true,
-        },
-    )
-    .expect("exact sni request failed");
-    run_h3_client_with_tls(
-        listen_addr,
-        H3TlsClientOptions {
-            server_name: "other.example.com",
-            authority: "other.example.com",
-            path: "/",
-            verify_peer: false,
-            root_cert_path: None,
-            client_identity: None,
-            application_protos: quiche::h3::APPLICATION_PROTOCOL,
-            send_request: true,
-        },
-    )
-    .expect("fallback sni request failed");
-    let metrics = rt
-        .block_on(scrape_metrics(
-            metrics_port,
-            "/metrics",
-            Duration::from_secs(20),
-        ))
-        .expect("metrics endpoint should become reachable");
-    assert!(
-        metrics.contains("spooky_downstream_tls_certificate_selection_total{listener=\"127.0.0.1:")
-    );
-    assert!(metrics.contains("selection=\"exact_sni\""));
-    assert!(metrics.contains("selection=\"fallback_unmatched_sni\""));
-    assert!(metrics.contains("spooky_downstream_tls_alpn_total"));
-    assert!(metrics.contains("protocol=\"h3\""));
-}
-
-#[test]
-#[serial_test::serial]
-fn bootstrap_tls_metrics_capture_missing_client_certificate_failures() {
-    if !local_listener_bind_available() {
-        return;
-    }
-    let dir = tempdir().expect("failed to create temp dir");
-    let (cert, key) = write_test_certs(&dir);
-    let (ca_cert, _client_cert, _client_key) =
-        write_test_ca_and_client_cert(&dir, "client-ca", "client.example.com");
-    let rt = tokio::runtime::Runtime::new().expect("runtime");
-    let backend_addr = rt.block_on(start_h2_backend());
-
-    let listen_port = find_free_tcp_port();
-    let metrics_port = find_free_tcp_port();
-    let mut config = make_config(
-        listen_port as u32,
-        backend_addr.to_string(),
-        cert.clone(),
-        key,
-    );
-    config.listen.tls.client_auth = ClientAuth {
-        enabled: true,
-        require_client_cert: true,
-        ca_file: Some(ca_cert),
-    };
-    config.observability.metrics.enabled = true;
-    config.observability.metrics.address = "127.0.0.1".to_string();
-    config.observability.metrics.port = metrics_port;
-    config.observability.metrics.path = "/metrics".to_string();
-
-    let _enter = rt.enter();
-    let listener = make_listener_with_bootstrap(config);
-    drop(_enter);
-    let listen_addr = listener.socket.local_addr().expect("listener addr");
-    let _listener_task = ListenerTaskGuard::spawn(&rt, listener);
-    let bootstrap_addr = SocketAddr::new(listen_addr.ip(), listen_port);
-
-    let err = rt
-        .block_on(run_bootstrap_h2_client_request_with_client_auth(
-            bootstrap_addr,
-            &cert,
-            "GET",
-            "/",
-            &[],
-            None,
-        ))
-        .expect_err("missing client cert should fail");
-    assert!(
-        err.contains("tls connect") || err.contains("sender ready") || err.contains("send request"),
-        "unexpected error: {err}"
-    );
-
-    let metrics = rt
-        .block_on(scrape_metrics(
-            metrics_port,
-            "/metrics",
-            Duration::from_secs(20),
-        ))
-        .expect("metrics endpoint should become reachable");
-    assert!(metrics.contains("spooky_downstream_tls_handshake_failure_total"));
-    assert!(metrics.contains("reason=\"missing_client_cert\""));
-}
-
-#[test]
-fn http3_head_suppresses_response_body() {
-    if !local_listener_bind_available() {
-        return;
-    }
-    let dir = tempdir().expect("failed to create temp dir");
-    let (cert, key) = write_test_certs(&dir);
-
-    let rt = tokio::runtime::Runtime::new().expect("runtime");
-    let backend_addr = rt.block_on(start_h2_backend_with_regression_routes());
-    let config = make_config(0, backend_addr.to_string(), cert, key);
-    let listener = QUICListener::new(config).expect("failed to create listener");
-    let listen_addr = listener.socket.local_addr().unwrap();
-    let _listener_task = ListenerTaskGuard::spawn(&rt, listener);
-
-    let response = run_h3_client_collect_response(
-        listen_addr,
-        vec![
-            quiche::h3::Header::new(b":method", b"HEAD"),
-            quiche::h3::Header::new(b":scheme", b"https"),
-            quiche::h3::Header::new(b":authority", b"localhost"),
-            quiche::h3::Header::new(b":path", b"/stream"),
-            quiche::h3::Header::new(b"user-agent", b"spooky-regression-test"),
-        ],
-        true,
-        Duration::from_secs(REQUEST_TIMEOUT_SECS),
-    )
-    .expect("HEAD request failed");
-
-    assert_eq!(response.status, "200");
-    assert!(
-        response.body.is_empty(),
-        "HEAD response must not include a body"
-    );
-    assert!(!response.reset, "HEAD response should complete cleanly");
-}
-
-#[test]
-fn bootstrap_h2_head_suppresses_response_body() {
-    if !local_listener_bind_available() {
-        return;
-    }
-    let dir = tempdir().expect("failed to create temp dir");
-    let (cert, key) = write_test_certs(&dir);
-
-    let rt = tokio::runtime::Runtime::new().expect("runtime");
-    let backend_addr = rt.block_on(start_h2_backend_with_regression_routes());
-    let listen_port = find_free_tcp_port();
-    let config = make_config(
-        u32::from(listen_port),
-        backend_addr.to_string(),
-        cert.clone(),
-        key,
-    );
-    let listener = make_listener_with_bootstrap(config);
-    let listen_addr = listener.socket.local_addr().unwrap();
-    let _listener_task = ListenerTaskGuard::spawn(&rt, listener);
-
-    let (status, body, _trailers) = rt
-        .block_on(run_bootstrap_h2_client_request(
-            SocketAddr::new(listen_addr.ip(), listen_port),
-            &cert,
-            "HEAD",
-            "/stream",
-            &[],
-        ))
-        .expect("bootstrap HEAD request failed");
-
-    assert_eq!(status, StatusCode::OK);
-    assert!(
-        body.is_empty(),
-        "bootstrap HEAD response must not include a body"
-    );
-}
-
-#[test]
-fn http3_rejects_upgrade_style_requests() {
-    if !local_listener_bind_available() {
-        return;
-    }
-    let dir = tempdir().expect("failed to create temp dir");
-    let (cert, key) = write_test_certs(&dir);
-
-    let rt = tokio::runtime::Runtime::new().expect("runtime");
-    let backend_addr = rt.block_on(start_h2_backend());
-    let config = make_config(0, backend_addr.to_string(), cert, key);
-    let listener = QUICListener::new(config).expect("failed to create listener");
-    let listen_addr = listener.socket.local_addr().unwrap();
-    let _listener_task = ListenerTaskGuard::spawn(&rt, listener);
-
-    let response = run_h3_client_collect_response(
-        listen_addr,
-        vec![
-            quiche::h3::Header::new(b":method", b"GET"),
-            quiche::h3::Header::new(b":scheme", b"https"),
-            quiche::h3::Header::new(b":authority", b"localhost"),
-            quiche::h3::Header::new(b":path", b"/"),
-            quiche::h3::Header::new(b"connection", b"Upgrade"),
-            quiche::h3::Header::new(b"upgrade", b"websocket"),
-        ],
-        true,
-        Duration::from_secs(REQUEST_TIMEOUT_SECS),
-    )
-    .expect("upgrade-style request failed");
-
-    assert_eq!(response.status, "400");
-    assert!(
-        String::from_utf8_lossy(&response.body).contains("Upgrade"),
-        "expected explicit Upgrade rejection body, got {:?}",
-        String::from_utf8_lossy(&response.body)
-    );
-    assert!(
-        !response.reset,
-        "upgrade rejection should be an HTTP response"
-    );
-}
-
-#[test]
-fn http3_to_http2_preserves_grpc_error_trailers() {
-    if !local_listener_bind_available() {
-        return;
-    }
-    let dir = tempdir().expect("failed to create temp dir");
-    let (cert, key) = write_test_certs(&dir);
-
-    let rt = tokio::runtime::Runtime::new().expect("runtime");
-    let backend_addr = rt.block_on(start_h2_backend_with_grpc_routes());
-    let config = make_config(0, backend_addr.to_string(), cert, key);
-    let listener = QUICListener::new(config).expect("failed to create listener");
-    let listen_addr = listener.socket.local_addr().unwrap();
-    let _listener_task = ListenerTaskGuard::spawn(&rt, listener);
-
-    let response = run_h3_client_collect_response(
-        listen_addr,
-        vec![
-            quiche::h3::Header::new(b":method", b"POST"),
-            quiche::h3::Header::new(b":scheme", b"https"),
-            quiche::h3::Header::new(b":authority", b"localhost"),
-            quiche::h3::Header::new(b":path", b"/grpc-error"),
-            quiche::h3::Header::new(b"content-type", b"application/grpc"),
-            quiche::h3::Header::new(b"content-length", b"0"),
-        ],
-        true,
-        Duration::from_secs(REQUEST_TIMEOUT_SECS),
-    )
-    .expect("grpc error request failed");
-
-    assert_eq!(response.status, "200");
-    assert!(
-        response.body.is_empty(),
-        "grpc error should be trailer-only"
-    );
-    assert!(
-        response
-            .trailers
-            .iter()
-            .any(|(name, value)| name.eq_ignore_ascii_case("grpc-status") && value == "14"),
-        "expected grpc-status=14 trailer, got {:?}",
-        response.trailers
-    );
-    assert!(
-        response.trailers.iter().any(
-            |(name, value)| name.eq_ignore_ascii_case("grpc-message") && value == "unavailable"
-        ),
-        "expected grpc-message trailer, got {:?}",
-        response.trailers
-    );
-}
-
-#[test]
-fn grpc_timeout_returns_recoverable_proxy_error() {
-    if !local_listener_bind_available() {
-        return;
-    }
-    let dir = tempdir().expect("failed to create temp dir");
-    let (cert, key) = write_test_certs(&dir);
-
-    let rt = tokio::runtime::Runtime::new().expect("runtime");
-    let backend_addr = rt.block_on(start_h2_backend_with_grpc_routes());
-    let mut config = make_config(0, backend_addr.to_string(), cert, key);
-    config.performance.backend_timeout_ms = 150;
-    let listener = QUICListener::new(config).expect("failed to create listener");
-    let listen_addr = listener.socket.local_addr().unwrap();
-    let _listener_task = ListenerTaskGuard::spawn(&rt, listener);
-
-    let response = run_h3_client_collect_response(
-        listen_addr,
-        vec![
-            quiche::h3::Header::new(b":method", b"POST"),
-            quiche::h3::Header::new(b":scheme", b"https"),
-            quiche::h3::Header::new(b":authority", b"localhost"),
-            quiche::h3::Header::new(b":path", b"/grpc-timeout"),
-            quiche::h3::Header::new(b"content-type", b"application/grpc"),
-            quiche::h3::Header::new(b"content-length", b"0"),
-        ],
-        true,
-        Duration::from_secs(REQUEST_TIMEOUT_SECS + 4),
-    )
-    .expect("grpc timeout request failed");
-
-    assert_eq!(response.status, "503");
-    assert!(
-        String::from_utf8_lossy(&response.body).contains("upstream timeout"),
-        "timeout body should explain the proxy failure"
-    );
-    assert!(
-        !response.reset,
-        "grpc timeout should surface as an HTTP response"
-    );
-}
-
-#[test]
-fn grpc_client_cancel_before_response_releases_stream() {
-    if !local_listener_bind_available() {
-        return;
-    }
-    let dir = tempdir().expect("tempdir");
-    let (cert, key) = write_test_certs(&dir);
-    let rt = tokio::runtime::Runtime::new().expect("runtime");
-
-    let backend_addr = rt.block_on(start_h2_backend_with_regression_routes());
-    let mut config = make_config(0, backend_addr.to_string(), cert, key);
-    config.performance.global_inflight_limit = 1;
-
-    let listener = QUICListener::new(config).expect("listener");
-    let listen_addr = listener.socket.local_addr().unwrap();
-    let _guard = ListenerTaskGuard::spawn(&rt, listener);
-
-    let (socket, local_addr, mut conn, mut h3) =
-        make_quic_client(listen_addr).expect("quic client");
-    let mut out = [0u8; MAX_UDP_PAYLOAD_BYTES];
-    let mut buf = [0u8; MAX_DATAGRAM_SIZE_BYTES];
-
-    let grpc_headers = vec![
-        quiche::h3::Header::new(b":method", b"POST"),
-        quiche::h3::Header::new(b":scheme", b"https"),
-        quiche::h3::Header::new(b":authority", b"localhost"),
-        quiche::h3::Header::new(b":path", b"/slow"),
-        quiche::h3::Header::new(b"content-type", b"application/grpc"),
-        quiche::h3::Header::new(b"content-length", b"0"),
-    ];
-    let stream_id = h3
-        .send_request(&mut conn, &grpc_headers, false)
-        .expect("send grpc request");
-
-    flush_quic(&mut conn, &socket, &mut out);
-    conn.stream_shutdown(stream_id, quiche::Shutdown::Write, 0)
-        .expect("stream shutdown");
-    flush_quic(&mut conn, &socket, &mut out);
-
-    let pump_end = Instant::now() + Duration::from_millis(200);
-    let mut events = Vec::new();
-    let mut io = PumpIo {
-        socket: &socket,
-        local_addr,
-        out: &mut out,
-        buf: &mut buf,
-    };
-    pump_h3_until(
-        &mut conn,
-        &mut h3,
-        &mut io,
-        stream_id,
-        pump_end,
-        &mut events,
-    );
-
-    let followup = run_h3_client_concurrent_get(
-        listen_addr,
-        &["/fast"],
-        Duration::from_secs(REQUEST_TIMEOUT_SECS + 4),
-    )
-    .expect("follow-up /fast request failed");
-    let fast = observation_for(&followup, "/fast");
-    assert_eq!(fast.status.as_deref(), Some("200"));
-}
+#[path = "h3_bridge/protocol.rs"]
+mod protocol;
 
 /// While draining, new QUIC Initial packets must be silently dropped so no new
 /// connection state is allocated.  The test starts draining before any client
